refactor: move the OSC 133 readers into shell/
What a shell tells the terminal about its own output, and what is built on top of
it: the ShellIntegration interface, the command-block and prompt-region scans, the
semantic block tracker behind SBQUERY, and folding.

These four belong together more strongly than anything else in the module.
CommandBlocks, PromptRegion and Folding are all backward walks over LOGICAL lines
behind a *LineSource dependency-injection seam -- the same shape three times, so
each can be exercised against a plain vector with no Grid, no Screen and no
terminal behind it -- and their doc comments already cross-reference each other by
name. Splitting them would have cost more than the directory is worth.

They keep that independence here, and a grep says so:

  grep '#include <vtbackend/grid/' src/vtbackend/shell/*.hpp

is empty. Two earlier commits are what bought it: LineFlags moved to core/, since
all three read it, and foldedRowsTopRow moved into Grid.hpp, since it was page
geometry that folding merely caused rather than anything about a fold.

FileUrl is NOT here despite arriving with the same shell-integration work. It is
OSC 7 and OSC 8, not OSC 133, and it went to core/ beside Hyperlink -- an OSC 8
target IS a URL, and FileUrl is how you read one.
