+++ src/vtbackend/Viewport.hpp
@@ -1,9 +1,9 @@
 // SPDX-License-Identifier: Apache-2.0
 #pragma once
 
-#include <vtbackend/Folding.hpp>
 #include <vtbackend/Screen.hpp>
 #include <vtbackend/core/Primitives.hpp>
+#include <vtbackend/shell/Folding.hpp>
 
 #include <crispy/LogStore.hpp>
