+++ src/vtbackend/shell/CommandBlocks_test.cpp
@@ -1,5 +1,5 @@
 // SPDX-License-Identifier: Apache-2.0
-#include <vtbackend/CommandBlocks.hpp>
+#include <vtbackend/shell/CommandBlocks.hpp>
 
 #include <catch2/catch_test_macros.hpp>
