+++ src/vtbackend/shell/Folding_test.cpp
@@ -1,8 +1,8 @@
 // SPDX-License-Identifier: Apache-2.0
-#include <vtbackend/Folding.hpp>
 #include <vtbackend/MockTerm.hpp>
 #include <vtbackend/StatusLineBuilder.hpp>
 #include <vtbackend/Terminal.hpp>
+#include <vtbackend/shell/Folding.hpp>
 
 #include <crispy/Utils.hpp>
