+++ src/vtbackend/shell/ShellIntegration.hpp
