+++ src/vtrasterizer/BoxDrawingRenderer_test.cpp
@@ -1,5 +1,5 @@
-#include <vtbackend/Folding.hpp>
 #include <vtbackend/core/Color.hpp>
+#include <vtbackend/shell/Folding.hpp>
 
 #include <vtrasterizer/BoxDrawingRenderer.hpp>
 #include <vtrasterizer/GridMetrics.hpp>
